convert story translation page to story page
